[Core] Move repository method from order to core
