+++ src/Sylius/Bundle/OrderBundle/Doctrine/ORM/OrderRepository.php
@@ -12,8 +12,6 @@
 namespace Sylius\Bundle\OrderBundle\Doctrine\ORM;
 
 use Sylius\Bundle\ResourceBundle\Doctrine\ORM\EntityRepository;
-use Sylius\Component\Core\OrderCheckoutStates;
-use Sylius\Component\Core\OrderPaymentStates;
 use Sylius\Component\Order\Model\OrderInterface;
 use Sylius\Component\Order\Repository\OrderRepositoryInterface;
 
@@ -105,25 +103,6 @@ class OrderRepository extends EntityRepository implements OrderRepositoryInterfa
             ->andWhere('o.state = :state')
             ->andWhere('o.updatedAt < :terminalDate')
             ->setParameter('state', OrderInterface::STATE_CART)
-            ->setParameter('terminalDate', $terminalDate)
-            ->getQuery()
-            ->getResult()
-        ;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function findOrdersUnpaidSince(\DateTime $terminalDate)
-    {
-        return $this->createQueryBuilder('o')
-            ->andWhere('o.checkoutState = :checkoutState')
-            ->andWhere('o.paymentState != :paymentState')
-            ->andWhere('o.state = :orderState')
-            ->andWhere('o.checkoutCompletedAt < :terminalDate')
-            ->setParameter('checkoutState', OrderCheckoutStates::STATE_COMPLETED)
-            ->setParameter('paymentState', OrderPaymentStates::STATE_PAID)
-            ->setParameter('orderState', OrderInterface::STATE_NEW)
             ->setParameter('terminalDate', $terminalDate)
             ->getQuery()
             ->getResult()
